Remove CheckReconcileCalled functions from the utils test pkg
These functions are very error-prone since most of our reconciliation
functions can be trigerred from multiple events that our tests don't
account for.
They are not used anymore except for the dynamic watches test: let's
move them to that package instead.
